requirements.txt
aiohttp
aiosignal
asgiref
async-timeout
attrs
certifi
charset-normalizer
click
colorama
datasets
dill
Django
django-cors-headers
djangorestframework
evaluate
filelock
frozenlist
fsspec
huggingface-hub
idna
joblib
MarkupSafe
mpmath
multidict
multiprocess
networkx
numpy
packaging
Pillow
pip
pyarrow
python-dateutil
pytz
PyYAML
requests
responses
safetensors
scikit-learn
scipy
sentence-transformers
sentencepiece
setfit
setuptools
six
sqlparse
sympy
tokenizers
torch
torchvision
tqdm
transformers
typing_extensions
tzdata
urllib3
xxhash
yarl